+++ src/Links/resources/database/migrations/2024_05_31_074502_add_root_item_to_link_group_items_table.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration {
+    public function up(): void
+    {
+        Schema::table('link_groups', function (Blueprint $table) {
+            $table->unsignedBigInteger('root_item_id')->nullable();
+
+            $table->foreign('root_item_id')->references('id')->on('link_group_items');
+        });
+    }
+
+    public function down(): void
+    {
+        Schema::table('link_groups', function (Blueprint $table) {
+            $table->dropColumn('root_item_id');
+        });
+    }
+};